Adding Exceptions to User for Testing

+++ LAB1/dst1_0626885/src/dst1/model/UserDAO.java
@@ -1,7 +1,6 @@
 package dst1.model;
 
 import java.util.List;
-
 import javax.persistence.*;
 
 public class UserDAO {
@@ -11,7 +10,6 @@ public class UserDAO {
 	
 	public UserDAO () {
 		this.entityManagerFactory = PersistenceUtil.getEntityManagerFactory();
-		
 	}
 	
 	/**
@@ -20,6 +18,10 @@ public class UserDAO {
 	 */
 	public void saveUser(User user) {
 		EntityManager entityManager = getEntityManager();
+		if(entityManager.getTransaction().isActive()) {
+			entityManager.persist(user);
+			return;
+		}
 		entityManager.getTransaction().begin();
 		entityManager.persist(user);
 		entityManager.getTransaction().commit();
@@ -47,6 +49,9 @@ public class UserDAO {
 	 */
 	public List<User> getAllUsers() {
 		EntityManager entityManager = getEntityManager();
+		if(entityManager.getTransaction().isActive())
+			return entityManager.createQuery( "from User", User.class ).getResultList();
+		
 		entityManager.getTransaction().begin();
 		List<User> result = entityManager.createQuery( "from User", User.class ).getResultList();
 		entityManager.getTransaction().commit();
@@ -57,12 +62,31 @@ public class UserDAO {
 	 * Remove a user from the current PersistenceContext
 	 * @param userID the userID of the user to be removed
 	 */
-	public void removeUser(Long userID) {
-		EntityManager entityManager = getEntityManager();
-		entityManager.getTransaction().begin();
-		User user = findUser(userID);
-		entityManager.remove(user);
-		entityManager.getTransaction().commit();
+	public void removeUser(Long userID) throws RemoveNullEntityException {
+		try {
+			EntityManager entityManager = getEntityManager();
+			if(entityManager.getTransaction().isActive()) {
+				User user = findUser(userID);
+				if(user == null) {
+					entityManager.getTransaction().rollback();
+					throw new RemoveNullEntityException("UserDAO: Trying to remove non-existent object");
+				}
+				
+				entityManager.remove(user);
+				return;
+			}
+			entityManager.getTransaction().begin();
+			User user = findUser(userID);
+			if(user == null) {
+				entityManager.getTransaction().rollback();
+				throw new RemoveNullEntityException("UserDAO: Trying to remove non-existent object");
+			}
+			
+			entityManager.remove(user);
+			entityManager.getTransaction().commit();
+		} catch(IllegalArgumentException iae) {
+			
+		}
 	}
 	
 	/**
